Removes v0 appointment transformers (#25018)
* Removes v0 appointment transformers


* Renames v2 appointments transformer


---------
